cheque passed

+++ src/main/java/peaksoft/services/impl/StopListServiceImpl.java
@@ -7,12 +7,13 @@ import peaksoft.dto.responses.SimpleResponse;
 import peaksoft.dto.responses.StopListResponse;
 import peaksoft.entity.MenuItem;
 import peaksoft.entity.StopList;
+import peaksoft.exeption.BadRequestException;
+import peaksoft.exeption.NotFoundException;
 import peaksoft.repositories.MenuItemRepository;
 import peaksoft.repositories.StopListRepository;
 import peaksoft.services.StopListService;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 /**
  * @author :ЛОКИ Kelsivbekov
@@ -30,15 +31,18 @@ public class StopListServiceImpl implements StopListService {
 
     @Override
     public SimpleResponse saveStopList(StopListRequest request, Long menuItemId) {
+        if (!menuItemRepository.existsById(menuItemId)){
+            throw new NotFoundException(String.format("Menu with Id: %d doesnt exist", menuItemId));
+        }
+
+
         List<StopList> all = stopListRepository.findAll();
         for (StopList stopList : all) {
             if (stopList.getDate().equals(request.date()))
-                return SimpleResponse.builder()
-                        .status(HttpStatus.BAD_REQUEST)
-                        .message("Бир дата менен бир тамак бир эле жолу сакталыш керек").build();
+                throw new BadRequestException("Бир дата менен бир тамак бир эле жолу сакталыш керек");
         }
 
-        MenuItem menuItem = menuItemRepository.findById(menuItemId).orElseThrow(() -> new NoSuchElementException(String.format("Menu with id: %d doesn't exist", menuItemId)));
+        MenuItem menuItem = menuItemRepository.findById(menuItemId).orElseThrow(() -> new NotFoundException(String.format("Menu with id: %d doesn't exist", menuItemId)));
 
         StopList stopList = new StopList();
         stopList.setReason(request.reason());
@@ -60,19 +64,24 @@ public class StopListServiceImpl implements StopListService {
 
     @Override
     public StopListResponse getById(Long id) {
+        if (!stopListRepository.existsById(id)){
+            throw new NotFoundException(String.format("Stop list with id: %d doesn't exist", id));
+        }
         return stopListRepository.findByIdStopList(id);
     }
 
     @Override
     public SimpleResponse update(Long id, StopListRequest request) {
+        if (!stopListRepository.existsById(id)){
+            throw new NotFoundException(String.format("Stop list with id: %d doesn't exist", id));
+        }
+
         StopList stopList = stopListRepository.findById(id).orElseThrow(
-                () -> new NoSuchElementException(String.format("Stop List with id: %d doesn't exist", id)));
+                () -> new NotFoundException(String.format("Stop List with id: %d doesn't exist", id)));
         List<StopList> all = stopListRepository.findAll();
         for (StopList stopList1 : all) {
             if (stopList1.getDate().equals(request.date()))
-                return SimpleResponse.builder()
-                        .status(HttpStatus.BAD_REQUEST)
-                        .message("Бир дата менен бир тамак бир эле жолу сакталыш керек").build();
+                throw new BadRequestException("Бир дата менен бир тамак бир эле жолу сакталыш керек");
         }
         stopList.setReason(request.reason());
         stopList.setDate(request.date());
@@ -86,16 +95,22 @@ public class StopListServiceImpl implements StopListService {
 
     @Override
     public SimpleResponse delete(Long menuId, Long stopListId) {
-        MenuItem menuItem = menuItemRepository.findById(menuId).orElseThrow(() -> new NoSuchElementException("Not found"));
-        StopList stopList = stopListRepository.findById(stopListId).orElseThrow(() -> new NoSuchElementException("Not found"));
+        if (!menuItemRepository.existsById(menuId)){
+            throw new NotFoundException(String.format("Menu with id: %d doesn't exist", menuId));
+        }
+
+        if (!stopListRepository.existsById(stopListId)){
+            throw new NotFoundException(String.format("Stop list with id: %d doesn't exist", stopListId));
+        }
+
+        MenuItem menuItem = menuItemRepository.findById(menuId).orElseThrow(() -> new NotFoundException(String.format("Menu with id: %d doesn't exist", menuId)));
+        StopList stopList = stopListRepository.findById(stopListId).orElseThrow(() -> new NotFoundException(String.format("Stop list with id: %d doesn't exist", stopListId)));
         if (stopList.getMenuItem().getId().equals(menuItem.getId())){
             menuItem.setStopList(null);
             menuItemRepository.save(menuItem);
             stopListRepository.delete(stopList);
         }else {
-                return SimpleResponse.builder()
-                        .status(HttpStatus.BAD_REQUEST)
-                        .message("Not found").build();
+                throw new BadRequestException("Not found");
             }
 
         return SimpleResponse.builder()
